refactor(bigquery): Refactor classes in external.rb to multiple files
